Move ViewModelCoroutineScope helper to ui:core
